adds core concepts and examples components
