+++ react-essentials/src/components/CoreConcepts.jsx
@@ -1,0 +1,14 @@
+import CoreConcept from './CoreConcept.jsx';
+import { CORE_CONCEPTS } from '../data.js';
+export default function () {
+  return (
+    <section id="core-concepts">
+      <h2>Core Concepts</h2>
+      <ul>
+        {CORE_CONCEPTS.map((conceptItem) => (
+          <CoreConcept key={conceptItem.title} {...conceptItem} />
+        ))}
+      </ul>
+    </section>
+  );
+}
