+++ react-essentials/src/components/Section.jsx
@@ -1,0 +1,8 @@
+export default function Section({ title, children }) {
+  return (
+    <section id={id}>
+      <h2>{title}</h2>
+      {children}
+    </section>
+  );
+}
